extracted auth panel ops
