--- client/src/components/AuthPanel/constants.ts
@@ -1,18 +1,0 @@
-export const AUTH_PANEL_OPTIONS = {
-  LOGIN: {
-    title: 'Member Login',
-    description: '',
-    submit: 'Submit',
-  },
-  RENEWAL: {
-    title: 'Membership Renewal',
-    description:
-      'Enter the login information that you used to create an account the last term that you were a member',
-    submit: 'Submit',
-  },
-  CHECK: {
-    title: 'Membership Check',
-    description: '',
-    submit: 'Check My Status',
-  },
-};
