+++ client/src/context/authPanel/state.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { /*type*/ AuthPanelOption } from 'components/AuthPanel/types';
 import { createContext } from 'react';
-import { AUTH_PANEL_OPTIONS } from 'components/AuthPanel/constants';
+import { AUTH_PANEL_OPTIONS } from 'utils/constants';
 
 export const initialState = {
   option: AUTH_PANEL_OPTIONS.LOGIN,
