+++ client/src/utils/constants.ts
@@ -2,3 +2,22 @@ export enum ROUTES {
   HOME = '/',
   MEMBERSHIP = '/membership',
 }
+
+export const AUTH_PANEL_OPTIONS = {
+  LOGIN: {
+    title: 'Member Login',
+    description: '',
+    submit: 'Submit',
+  },
+  RENEWAL: {
+    title: 'Membership Renewal',
+    description:
+      'Enter the login information that you used to create an account the last term that you were a member',
+    submit: 'Submit',
+  },
+  CHECK: {
+    title: 'Membership Check',
+    description: '',
+    submit: 'Check My Status',
+  },
+};
